apps/canary-atomy-tui/requirements.txt
textual[dev]
httpx
pydantic
python-dotenv
